requirements.txt
amqp==5.3.1
appdirs==1.4.4
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
bs4==0.0.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssselect==1.2.0
Django==5.1.5
django-neomodel==0.0.7
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docopt==0.6.2
fake-useragent==2.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
kombu==5.4.2
lxml==5.3.0
neo4j==5.15.0
neo4j-driver==5.15.0
neomodel==5.2.1
outcome==1.3.0.post0
owlready2==0.47
packaging==24.2
pipreqs==0.4.13
prompt_toolkit==3.0.50
pycparser==2.22
pyee==11.1.1
PyJWT==2.10.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
rdflib==7.1.4
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.29.0
selenium-stealth==1.0.6
setuptools==78.0.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.5.5
urllib3==1.26.20
vine==5.1.0
w3lib==2.3.1
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarg==0.1.10
zipp==3.21.0
